Remove more dead code. Include semantic layer logic...
This layer is incompatible with the SQL approach we're presently using for transformations.

That's okay. We can hand-tune transformations in configurations for now, and address
semantics once we have a GUI to build. It could leverage SQL, or not, but for now
it's too complex and not helpful enough to justify completing the implementation.