Effect workspaces (#10)
* add script to pull gbif schema and reformat as json object
* pull required field list from github and merge with schema
* add integer validator
* allow specifying full path, including file name, to the config file.
* populate validators in schema
* add rocky intertidal sub repo for testing
* switch to using json schema

change schema generation to only make fields required in primary table. ie not event and occurance but rather just event.
move validation profile setting into each dataset in the schema so that we can use a different profile for each dataset

* Add foreign keys and Export db

Add foreign key constrains to db tables
Option to export db to file at the end of the transformation
move output config into it's own section for transformations
Add option to strip null columns from csv output

* Derive types from data and schema objects and remove obsolete types (#6)
* Simplify types by deriving from schemas more consistently
* Add some basic PR safeguards, remove unused agents from .claude
* add example github action to run cli
* add junit report summary to test output in github action
* Run all code quality actions in one instance
* Take guess-work out of validation violations.

Removes ad-hoc object creation and type-casting.
Also removes use of arbitrary text-based unions.
* Remove more dead code. Include semantic layer logic...

This layer is incompatible with the SQL approach we're presently using for transformations.

That's okay. We can hand-tune transformations in configurations for now, and address
semantics once we have a GUI to build. It could leverage SQL, or not, but for now
it's too complex and not helpful enough to justify completing the implementation.
* re-add code to some error classes. move csv creation code into shared function and use in transform and validation
* fix schema profile structure. datasets should be under validation
* add database schema table creation to validation so that it attempts to insert csv data into the defined schema tables, the same as transformation. This will take care of a large section of validation for us.
* Update CLAUDE.md to reflect recent changes
* Add row-level validation to new duckdb validation convention

---------

Co-authored-by: Matthew Foster <matthew.foster@hakai.org>